+++ src/database.rs
@@ -4,8 +4,9 @@ use mongodb::{
     Client as MongoClient,
     Collection, Database, bson::doc
 };
+use serenity::prelude::Context;
 
-use crate::models::guild::Guild;
+use crate::{models::guild::Guild, context::context_get_guild};
 
 pub async fn get_rexmit_database() -> Option<Database> {
     let database_url: Result<String, env::VarError> = std::env::var("DATABASE_URL");
@@ -88,4 +89,37 @@ pub async fn clear_guild_queue(guild: serenity::model::prelude::Guild) {
             }
         }
     }
+}
+
+pub async fn set_joined(ctx: &Context, guild_id: u64, joined: bool) -> bool {
+    let collection_option = get_guild_collection().await;
+    if collection_option.is_some() {
+        let collection = collection_option.unwrap();
+        let partial_guild_option = context_get_guild(ctx, guild_id.into()).await;
+        if partial_guild_option.is_some() {
+            let guild = Guild::new_from_serenity_partial_guild(partial_guild_option);
+            let result = collection.find_one_and_update(doc! { "id": &guild_id.to_string() }, doc! { "$set": { "joined": joined }}, None).await;
+    
+            println!("{:?}", result);
+            match &result {
+                Ok(option) => {
+                    match &option {
+                        Some(guild) => {
+                            println!("{:?}", guild);
+                        }, 
+                        None => {
+                            let result = collection.insert_one(&guild, None).await;
+                            println!("{:?}", result);
+                        }
+                    }
+                },
+                Err(why) => {
+                    println!("{}", why);
+                }
+            }
+            return true;
+        }
+    }
+    println!("unable to set joined status");
+    return false;
 }
