+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod models;
 pub mod command;
 pub mod database;
+pub mod context;
